requirements.txt
discord.py
python-dotenv
graphviz
emoji
